# docs/swagger.yaml
basePath: /api/public/accounting/v1
definitions:
...
  dto.CreateAccountGroupRequest:
    properties:
      account_nature:
        enum:
        - ASSET
        - LIABILITY
        - EQUITY
        - INCOME
        - EXPENSE
        example: ASSET
        type: string
      book_id:
        example: 550e8400-e29b-41d4-a716-446655440010
        type: string
      code:
        example: CA
        maxLength: 30
        type: string
      is_system:
        example: false
        type: boolean
      name:
        example: Current Assets
        maxLength: 200
        type: string
      parent_group_id:
        example: 550e8400-e29b-41d4-a716-446655440040
        type: string
      sort_order:
        example: 10
        minimum: 0
        type: integer
    required:
    - account_nature
    - book_id
    - code
    - name
    - sort_order
    type: object
  dto.CreateAccountRequest:
    properties:
      account_nature:
        enum:
        - ASSET
        - LIABILITY
        - EQUITY
        - INCOME
        - EXPENSE
        example: ASSET
        type: string
      allow_manual_posting:
        example: true
        type: boolean
      book_id:
        example: 550e8400-e29b-41d4-a716-446655440010
        type: string
      code:
        example: "110010"
        maxLength: 50
        type: string
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      control_type:
        example: CASH
        maxLength: 50
        type: string
      display_name:
        example: Cash
        maxLength: 200
        type: string
      group_id:
        example: 550e8400-e29b-41d4-a716-446655440050
        type: string
      is_active:
        example: true
        type: boolean
      is_system:
        example: false
        type: boolean
      name:
        example: Cash on Hand
        maxLength: 200
        type: string
      normal_balance:
        enum:
        - D
        - C
        example: D
        type: string
      parent_account_id:
        example: 550e8400-e29b-41d4-a716-446655440060
        type: string
      require_branch:
        example: false
        type: boolean
      require_cost_center:
        example: false
        type: boolean
      require_employee:
        example: false
        type: boolean
      require_party:
        example: false
        type: boolean
      require_tax_breakup:
        example: false
        type: boolean
      usage:
        enum:
        - HEADER
        - POSTABLE
        example: POSTABLE
        type: string
    required:
    - account_nature
    - book_id
    - code
    - company_id
    - group_id
    - name
    - normal_balance
    - usage
    type: object
...
  dto.UpdateAccountGroupRequest:
    properties:
      name:
        example: Current Assets (Operating)
        maxLength: 200
        type: string
      parent_group_id:
        example: 550e8400-e29b-41d4-a716-446655440040
        type: string
      sort_order:
        example: 15
        minimum: 0
        type: integer
    type: object
  dto.UpdateAccountRequest:
    properties:
      allow_manual_posting:
        example: true
        type: boolean
      control_type:
        example: CASH
        maxLength: 50
        type: string
      display_name:
        example: Petty Cash
        maxLength: 200
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Cash on Hand — Main
        maxLength: 200
        type: string
      parent_account_id:
        example: 550e8400-e29b-41d4-a716-446655440060
        type: string
      require_branch:
        example: false
        type: boolean
      require_cost_center:
        example: false
        type: boolean
      require_employee:
        example: false
        type: boolean
      require_party:
        example: false
        type: boolean
      require_tax_breakup:
        example: false
        type: boolean
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        items:
          type: string
        type: array
      message:
        type: string
      status:
        type: boolean
      status_code:
        type: integer
    type: object
host: localhost:8086
info:
  contact:
    email: support@queuebuster.co
    name: API Support
    url: http://www.queuebuster.co/support
  description: Enterprise accounting module for QueueBuster ERP
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: QB Accounting Service API
  version: 1.0.0
paths:
  /account-groups:
    post:
      consumes:
      - application/json
      description: Create a new account group under a book
      parameters:
      - description: Account group data
        in: body
        name: account_group
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccountGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create an account group
      tags:
      - account-groups
  /account-groups/{id}:
    delete:
      description: Delete an account group
      parameters:
      - description: Account group ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete an account group
      tags:
      - account-groups
    get:
      description: Get an account group by ID
      parameters:
      - description: Account group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get an account group
      tags:
      - account-groups
    put:
      consumes:
      - application/json
      description: Update an account group
      parameters:
      - description: Account group ID
        in: path
        name: id
        required: true
        type: string
      - description: Account group fields
        in: body
        name: account_group
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccountGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update an account group
      tags:
      - account-groups
  /account-groups/book/{book_id}:
    get:
      description: List account groups for a book
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List account groups by book
      tags:
      - account-groups
...
  /accounts:
    post:
      consumes:
      - application/json
      description: Create a new ledger account
      parameters:
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create an account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      description: Delete an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete an account
      tags:
      - accounts
    get:
      description: Get an account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get an account
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Update an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account fields
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update an account
      tags:
      - accounts
  /accounts/book/{book_id}:
    get:
      description: List accounts for an accounting book
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List accounts by book
      tags:
      - accounts
  /accounts/company/{company_id}:
    get:
      description: List accounts for a company
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List accounts by company
      tags:
      - accounts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"